Update all snaps-execution-environments tests to use Webdriverio
