+++ packages/snaps-execution-environments/src/common/endowments/network.test.ts
@@ -1,8 +1,16 @@
+/**
+ * @jest-environment jsdom
+ */
+
 import fetchMock from 'jest-fetch-mock';
 
 import network from './network';
 
 describe('Network endowments', () => {
+  beforeAll(() => {
+    globalThis.harden = (value) => value;
+  });
+
   describe('fetch', () => {
     beforeEach(() => {
       fetchMock.enableMocks();
