+++ packages/snaps-utils/src/test-utils/index.ts
@@ -8,4 +8,5 @@ export * from './common';
 export * from './manifest';
 export * from './server';
 export * from './snap';
+export * from './spy';
 export * from './stream';
